Create separate planner package
